requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
aiohappyeyeballs=2.4.4=py39hecd8cb5_0
aiohttp=3.11.10=py39h46256e1_0
aiosignal=1.2.0=pyhd3eb1b0_0
annotated-types=0.6.0=py39hecd8cb5_0
ansi2html=1.9.1=py39hecd8cb5_0
anyio=4.6.2=py39hecd8cb5_0
appnope=0.1.2=py39hecd8cb5_1001
asgiref=3.8.1=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
async-timeout=4.0.3=py39hecd8cb5_0
attrs=24.3.0=py39hecd8cb5_0
backcall=0.2.0=pyhd3eb1b0_0
backoff=2.2.1=pypi_0
bcrypt=4.2.1=pypi_0
blas=1.0=mkl
blinker=1.6.2=py39hecd8cb5_0
bokeh=2.4.3=py39hecd8cb5_0
bottleneck=1.4.2=py39hf3fd67a_0
brotli=1.0.9=h6c40b1e_8
brotli-bin=1.0.9=h6c40b1e_8
brotli-python=1.0.9=py39hcec6c5f_8
build=1.2.2.post1=pypi_0
ca-certificates=2024.11.26=hecd8cb5_0
cachetools=5.5.0=pypi_0
certifi=2024.12.14=py39hecd8cb5_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
chroma-hnswlib=0.7.6=pypi_0
chromadb=0.5.23=pypi_0
click=8.1.7=py39hecd8cb5_0
cloudpickle=3.0.0=py39hecd8cb5_0
coloredlogs=15.0.1=pypi_0
comm=0.2.1=py39hecd8cb5_0
contourpy=1.2.0=py39ha357a0b_0
cycler=0.11.0=pyhd3eb1b0_0
cytoolz=0.12.2=py39h6c40b1e_0
dash=2.14.2=py39hecd8cb5_0
dask=2023.4.1=py39hecd8cb5_1
dask-core=2023.4.1=py39hecd8cb5_0
dataclasses-json=0.6.5=py39hecd8cb5_0
debugpy=1.6.7=py39hcec6c5f_0
decorator=5.1.1=pyhd3eb1b0_0
deprecated=1.2.15=pypi_0
distributed=2023.4.1=py39hecd8cb5_1
distro=1.9.0=py39hecd8cb5_0
durationpy=0.9=pypi_0
et_xmlfile=1.1.0=py39hecd8cb5_0
exceptiongroup=1.2.0=py39hecd8cb5_0
executing=0.8.3=pyhd3eb1b0_0
fastapi=0.115.6=pypi_0
filelock=3.16.1=pypi_0
flask=3.0.3=py39hecd8cb5_0
flask-compress=1.13=py39hecd8cb5_0
flatbuffers=24.12.23=pypi_0
fonttools=4.51.0=py39h6c40b1e_0
freetype=2.12.1=hd8bbffd_0
frozenlist=1.5.0=py39h46256e1_0
fsspec=2024.6.1=py39hecd8cb5_0
google-auth=2.37.0=pypi_0
googleapis-common-protos=1.66.0=pypi_0
greenlet=3.1.1=py39h6d0c2b6_0
groq=0.13.1=pypi_0
grpcio=1.68.1=pypi_0
h11=0.14.0=py39hecd8cb5_0
heapdict=1.0.1=pyhd3eb1b0_0
httpcore=1.0.2=py39hecd8cb5_0
httptools=0.6.4=pypi_0
httpx=0.27.0=py39hecd8cb5_0
httpx-sse=0.4.0=py39hecd8cb5_0
huggingface-hub=0.27.0=pypi_0
humanfriendly=10.0=pypi_0
idna=3.7=py39hecd8cb5_0
importlib-metadata=8.5.0=py39hecd8cb5_0
importlib_metadata=8.5.0=hd3eb1b0_0
importlib_resources=6.4.0=py39hecd8cb5_0
intel-openmp=2023.1.0=ha357a0b_43548
ipykernel=6.29.5=py39hecd8cb5_0
ipython=8.15.0=py39hecd8cb5_0
itsdangerous=2.2.0=py39hecd8cb5_0
jedi=0.19.2=py39hecd8cb5_0
jinja2=3.1.4=py39hecd8cb5_1
jiter=0.6.1=py39hcc58cb1_0
jpeg=9e=h46256e1_3
jsonpatch=1.33=py39hecd8cb5_1
jsonpointer=2.1=pyhd3eb1b0_0
jupyter_client=8.6.0=py39hecd8cb5_0
jupyter_core=5.7.2=py39hecd8cb5_0
kiwisolver=1.4.4=py39hcec6c5f_0
kubernetes=31.0.0=pypi_0
langchain=0.3.13=pypi_0
langchain-community=0.3.12=py39hecd8cb5_0
langchain-core=0.3.28=pypi_0
langchain-experimental=0.3.4=pypi_0
langchain-groq=0.2.2=pypi_0
langchain-openai=0.2.14=pypi_0
langchain-text-splitters=0.3.3=py39hecd8cb5_0
langgraph=0.2.60=pypi_0
langgraph-checkpoint=2.0.9=pypi_0
langgraph-sdk=0.1.48=pypi_0
langsmith=0.1.147=py39hecd8cb5_0
lcms2=2.16=h4f63f0c_0
lerc=4.0.0=h6d0c2b6_0
libbrotlicommon=1.0.9=h6c40b1e_8
libbrotlidec=1.0.9=h6c40b1e_8
libbrotlienc=1.0.9=h6c40b1e_8
libcxx=14.0.6=h9765a3e_0
libdeflate=1.22=h46256e1_0
libffi=3.4.4=hecd8cb5_1
libpng=1.6.39=h6c40b1e_0
libsodium=1.0.18=h1de35cc_0
libtiff=4.5.1=h6fa9cd1_1
libwebp-base=1.3.2=h46256e1_1
llama-stack=0.0.1a5=pypi_0
locket=1.0.0=py39hecd8cb5_0
lz4-c=1.9.4=hcec6c5f_1
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=py39h6c40b1e_0
marshmallow=3.19.0=py39hecd8cb5_0
matplotlib=3.9.2=py39hecd8cb5_1
matplotlib-base=3.9.2=py39h919b35b_1
matplotlib-inline=0.1.6=py39hecd8cb5_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h8e150cf_43560
mkl-service=2.4.0=py39h6c40b1e_1
mkl_fft=1.3.8=py39h6c40b1e_0
mkl_random=1.2.4=py39ha357a0b_0
mmh3=5.0.1=pypi_0
monotonic=1.6=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.1.0=pypi_0
multidict=6.1.0=py39h46256e1_0
mypy_extensions=1.0.0=py39hecd8cb5_0
ncurses=6.4=hcec6c5f_0
nest-asyncio=1.6.0=py39hecd8cb5_0
numexpr=2.8.7=py39h827a554_0
numpy=1.26.4=py39h827a554_0
numpy-base=1.26.4=py39ha186be2_0
oauthlib=3.2.2=pypi_0
onnxruntime=1.19.2=pypi_0
openai=1.58.1=pypi_0
openjpeg=2.5.2=hbf2204d_0
openpyxl=3.1.5=py39h46256e1_0
openssl=3.0.15=h46256e1_0
opentelemetry-api=1.29.0=pypi_0
opentelemetry-exporter-otlp-proto-common=1.29.0=pypi_0
opentelemetry-exporter-otlp-proto-grpc=1.29.0=pypi_0
opentelemetry-instrumentation=0.50b0=pypi_0
opentelemetry-instrumentation-asgi=0.50b0=pypi_0
opentelemetry-instrumentation-fastapi=0.50b0=pypi_0
opentelemetry-proto=1.29.0=pypi_0
opentelemetry-sdk=1.29.0=pypi_0
opentelemetry-semantic-conventions=0.50b0=pypi_0
opentelemetry-util-http=0.50b0=pypi_0
orjson=3.10.7=py39h7854f77_0
overrides=7.7.0=pypi_0
packaging=24.2=py39hecd8cb5_0
pandas=2.2.3=py39h6d0c2b6_0
parso=0.8.4=py39hecd8cb5_0
partd=1.4.1=py39hecd8cb5_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=11.0.0=py39h47bf62f_1
pip=24.2=py39hecd8cb5_0
platformdirs=3.10.0=py39hecd8cb5_0
plotly=5.24.1=py39h20db666_0
posthog=3.7.4=pypi_0
prompt-toolkit=3.0.43=py39hecd8cb5_0
propcache=0.2.0=py39h46256e1_0
protobuf=5.29.2=pypi_0
psutil=5.9.0=py39hca72f7f_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pydantic=2.10.3=py39hecd8cb5_0
pydantic-core=2.27.1=py39h83de92b_0
pydantic-settings=2.6.1=py39hecd8cb5_0
pygments=2.15.1=py39hecd8cb5_1
pyparsing=3.2.0=py39hecd8cb5_0
pypika=0.48.9=pypi_0
pyproject-hooks=1.2.0=pypi_0
pysocks=1.7.1=py39hecd8cb5_0
python=3.9.21=hce00570_1
python-dateutil=2.9.0post0=py39hecd8cb5_2
python-dotenv=0.21.0=py39hecd8cb5_0
python-lmdb=1.4.1=py39hcec6c5f_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytz=2024.1=py39hecd8cb5_0
pyyaml=6.0.2=py39h46256e1_0
pyzmq=26.2.0=py39h6d0c2b6_0
readline=8.2=hca72f7f_0
regex=2024.11.6=pypi_0
requests=2.32.3=py39hecd8cb5_1
requests-oauthlib=2.0.0=pypi_0
requests-toolbelt=1.0.0=py39hecd8cb5_0
retrying=1.3.3=pyhd3eb1b0_2
rich=13.9.4=pypi_0
rsa=4.9=pypi_0
seaborn=0.13.2=py39hecd8cb5_0
setuptools=75.1.0=py39hecd8cb5_0
shellingham=1.5.4=pypi_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.3.0=py39hecd8cb5_0
sortedcontainers=2.4.0=pyhd3eb1b0_0
sqlalchemy=2.0.34=py39hd53568b_0
sqlite=3.45.3=h6c40b1e_0
stack_data=0.2.0=pyhd3eb1b0_0
starlette=0.41.3=pypi_0
sympy=1.13.3=pypi_0
tabulate=0.9.0=py39hecd8cb5_0
tbb=2021.8.0=ha357a0b_0
tblib=1.7.0=pyhd3eb1b0_0
tenacity=9.0.0=py39hecd8cb5_0
tiktoken=0.8.0=pypi_0
tk=8.6.14=h4d00af3_0
tokenizers=0.20.3=pypi_0
tomli=2.2.1=pypi_0
toolz=0.12.0=py39hecd8cb5_0
tornado=6.4.2=py39h46256e1_0
tqdm=4.66.5=py39h20db666_0
traitlets=5.14.3=py39hecd8cb5_0
typer=0.15.1=pypi_0
typing-extensions=4.12.2=py39hecd8cb5_0
typing_extensions=4.12.2=py39hecd8cb5_0
typing_inspect=0.9.0=py39hecd8cb5_0
tzdata=2024b=h04d1e81_0
unicodedata2=15.1.0=py39h6c40b1e_0
urllib3=2.2.3=py39hecd8cb5_0
uvicorn=0.34.0=pypi_0
uvloop=0.21.0=pypi_0
watchfiles=1.0.3=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
websocket-client=1.8.0=pypi_0
websockets=14.1=pypi_0
werkzeug=3.0.6=py39hecd8cb5_0
wheel=0.44.0=py39hecd8cb5_0
wrapt=1.17.0=pypi_0
xz=5.4.6=h6c40b1e_1
yaml=0.2.5=haf1e3a3_0
yarl=1.18.0=py39h46256e1_0
zeromq=4.3.5=hcec6c5f_0
zict=3.0.0=py39hecd8cb5_0
zipp=3.21.0=py39hecd8cb5_0
zlib=1.2.13=h4b97444_1
zstd=1.5.6=h138b38a_0
